refactor sqlite tests
